+++ src/main/java/com/chungnamthon/cheonon/receipt_ocr/domain/ReceiptPreview.java
@@ -1,13 +1,15 @@
 package com.chungnamthon.cheonon.receipt_ocr.domain;
 
-import com.chungnamthon.cheonon.map.domain.Merchant;
+import com.chungnamthon.cheonon.local_merchant.domain.Merchant;
 import com.chungnamthon.cheonon.user.domain.User;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.LocalTime;
 
 @Entity
 @Table(name = "receipt_preview")
@@ -29,8 +31,11 @@ public class ReceiptPreview {
     @JoinColumn(name = "merchant_id", nullable = false)
     private Merchant merchant;
 
-    @Column(name = "visit_datetime", nullable = false)
-    private LocalDateTime visitDateTime;
+    @Column(name = "visit_date", nullable = false)
+    private LocalDate visitDate;
+
+    @Column(name = "visit_time", nullable = false)
+    private LocalTime visitTime;
 
     @Column(updatable = false)
     private LocalDateTime createdAt = LocalDateTime.now();
