+++ src/main/java/com/chungnamthon/cheonon/receipt_ocr/dto/ReceiptConfirmResponseDto.java
@@ -3,7 +3,9 @@ package com.chungnamthon.cheonon.receipt_ocr.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.LocalTime;
 
 @Getter
 @Builder
@@ -12,7 +14,8 @@ public class ReceiptConfirmResponseDto {
     private Long          userId;
     private Long          merchantId;
     private Long          pointId;
-    private LocalDateTime visitDateTime;
+    private LocalDate visitDate;
+    private LocalTime visitTime;
     private LocalDateTime createdAt;
     private Integer       currentPoint;
 }
